сохранение игрового поля и действия

+++ scripts/TraderAbility/TraderAbility.gml
@@ -13,7 +13,7 @@ function TraderAbility(using_figure, using_cell) : FigureAbilityAction() constru
 	
 	change_summon_button = function() {
 		O_SummonButton.back = 0;
-		O_SummonButton.y = O_SummonButton.standart_y + 55;
+		O_SummonButton.y = O_SummonButton.standart_y + 75;
 		O_SummonButton.change_sprite(O_SummonButton.standart_sprite, 0.2);
 		O_SummonButton.image_index = 2
 	}
@@ -29,6 +29,7 @@ function TraderAbility(using_figure, using_cell) : FigureAbilityAction() constru
 			buttons[i] = new_button
 			figure_button_x += 130;
 		}
+		
 		O_App.data.save(using_figure.owner, load_data);
 	}
 	if global.using_ability {
@@ -41,19 +42,24 @@ function TraderAbility(using_figure, using_cell) : FigureAbilityAction() constru
 		target_cell.create_figure(chosen_button.figure_type);
 		instance_destroy(chosen_button)
 		for (i = 0; i < 3; i++) {
-			new_figure = instance_create_depth(buttons[i].x, buttons[i].y, 1, O_Figure);
-			new_figure.set_behaviour(buttons[i].figure_type);
-			show_debug_message(buttons[i].figure_type)
-			new_figure.drop();
-			instance_destroy(buttons[i]);
+			if instance_exists(buttons[i]) {
+				new_figure = instance_create_depth(buttons[i].x, buttons[i].y, 1, O_Figure);
+				new_figure.set_behaviour(buttons[i].figure_type);
+				new_figure.drop();
+				instance_destroy(buttons[i]);
+			}
 		}
 	}
-	
+
 	check_ability_targets = function(a, b) {
 		O_GameField.check_controlled_summon_cells(global.turn_owner);
 		load_data = O_App.data.load(global.turn_owner);
 		if array_length(load_data.player_figures) < 3 and !global.using_ability{
-			O_GameField.clear_all_marks();	
+			O_GameField.clear_all_marks();
+		}
+		if global.using_ability and instance_exists(O_AbilityInputController){
+			O_AbilityInputController.start_ability();
+			O_GameField.clear_all_marks();
 		}
 	}
 	
@@ -66,13 +72,12 @@ function TraderAbility(using_figure, using_cell) : FigureAbilityAction() constru
 	
 	global.cell_action = function(cell) {
 		if (cell.marked) {
+			global.cell_click_callback.set_draw_marks(1);
 			global.cell_click_callback = cell;
-			if !O_GameLoopController.have_action() {O_AbilityInputController.start_ability()}
 			if O_GameLoopController.action.chosen_button != undefined {
 				O_EndTurn.active = 1
 				cell.set_draw_marks(0)
 				}
-			cell.set_draw_marks(1)
 			cell.set_draw_marks(0)
 			O_GameLoopController.action.target_cell = cell;
 			O_EndTurn.unblock();
@@ -88,12 +93,12 @@ function TraderAbility(using_figure, using_cell) : FigureAbilityAction() constru
 		O_SummonButton.change_sprite(S_Back, 0.2);
 		O_SummonButton.image_index = 0;
 		O_SummonButton.back = 1;
+		if target_cell == undefined{check_ability_targets(1, 1);}
 	}
 	
 	back = function() {
 		if chosen_button != undefined or target_cell != undefined{
 			chosen_button.image_alpha = 1;
-			O_EndTurn.block();
 			chosen_button = undefined;
 			sprite_draw = undefined;
 			target_cell = undefined;
@@ -101,6 +106,8 @@ function TraderAbility(using_figure, using_cell) : FigureAbilityAction() constru
 			O_SummonButton.change_sprite(O_SummonButton.standart_sprite, 0.2);
 			O_SummonButton.image_index = 2;
 			O_SummonButton.back = 0;
+			O_EndTurn.block();
+			O_GameField.clear_all_marks();
 		}
 	}
